+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: kazumanoda <kazumanoda@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/27 23:29:17 by knoda             #+#    #+#             */
-/*   Updated: 2020/07/05 19:25:57 by kazumanoda       ###   ########.fr       */
+/*   Updated: 2020/07/10 21:58:50 by kazumanoda       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ char	*ft_strnstr(const char *big, const char *little, size_t len)
 {
 	if (!*little)
 		return ((char	*)big);
-	while (*big && (int)len >= ft_strlen(little))
+	while (*big && len >= ft_strlen(little))
 	{
 		if (ft_strncmp(big, little, ft_strlen(little)) == 0)
 			return ((char	*)big);
