+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: kazumanoda <kazumanoda@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/27 23:30:27 by knoda             #+#    #+#             */
-/*   Updated: 2020/06/30 01:31:45 by kazumanoda       ###   ########.fr       */
+/*   Updated: 2020/07/10 21:28:42 by kazumanoda       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void			*ft_memccpy(void *dest, const void *src, int c, size_t n);
 void			*ft_memmove(void *buf1, const void *buf2, size_t n);
 void			*ft_memchr(const void *buf, int ch, size_t n);
 int				ft_memcmp(const void *buf1, const void *buf2, size_t n);
-int				ft_strlen(const char *str);
+size_t			ft_strlen(const char *str);
 size_t			ft_strlcpy(char *dest, const char *src, size_t size);
 size_t			ft_strlcat(char *s1, const char *s2, size_t size);
 char			*ft_strchr(const char *s, int c);
